+++ tests/Feature/Stimulus/StimulusProviderTest.php
@@ -13,6 +13,7 @@ use Symfony\Lsp\Feature\Stimulus\StimulusCompletionContextResolver;
 use Symfony\Lsp\Feature\Stimulus\StimulusCompletionProvider;
 use Symfony\Lsp\Feature\Stimulus\StimulusController;
 use Symfony\Lsp\Feature\Stimulus\StimulusControllerExtractor;
+use Symfony\Lsp\Feature\Stimulus\StimulusControllerNameNormalizer;
 use Symfony\Lsp\Feature\Stimulus\StimulusDiagnosticProvider;
 use Symfony\Lsp\Feature\Stimulus\StimulusDocumentLinkProvider;
 use Symfony\Lsp\Feature\Stimulus\StimulusExtractor;
@@ -37,10 +38,11 @@ final class StimulusProviderTest extends TestCase
         $converter = new PositionConverter();
         $comments = new TwigCommentParser();
         $codeMasker = new JavaScriptSourceAnalyzer();
+        $controllerNameNormalizer = new StimulusControllerNameNormalizer();
         $extractor = new StimulusExtractor(
-            new StimulusControllerExtractor($converter, new ProjectPathResolver(new UriToPathConverter()), $codeMasker),
-            new StimulusReferenceExtractor($converter, $comments, $codeMasker),
-            new StimulusCompletionContextResolver($converter, $comments),
+            new StimulusControllerExtractor($converter, new ProjectPathResolver(new UriToPathConverter()), $codeMasker, $controllerNameNormalizer),
+            new StimulusReferenceExtractor($converter, $comments, $codeMasker, $controllerNameNormalizer),
+            new StimulusCompletionContextResolver($converter, $comments, $controllerNameNormalizer),
         );
         $controllerUri = 'file:///workspace/assets/controllers/search_controller.js';
         $controllerText = <<<'JS'
@@ -71,6 +73,11 @@ final class StimulusProviderTest extends TestCase
             <div data-controller="admin-{{ dataController }}"></div>
             <div data-controller="{{ dataController }}-admin"></div>
             {{ stimulus_action('search', 'open') }}
+            {{ stimulus_controller('symfony/ux-autocomplete/autocomplete') }}
+            {{ stimulus_controller('@symfony/ux-autocomplete/autocomplete') }}
+            {{ stimulus_controller('symfony--ux-autocomplete--autocomplete') }}
+            {{ stimulus_action('symfony/ux-autocomplete/autocomplete', 'onChange') }}
+            {{ stimulus_target('symfony/ux-autocomplete/autocomplete', 'field') }}
             TWIG;
         $documents = new DocumentStore();
         $documents->open(new Document($controllerUri, 'javascript', 1, $controllerText));
@@ -79,29 +86,52 @@ final class StimulusProviderTest extends TestCase
         $controllerCompletionUri = 'file:///workspace/templates/controller_completion.html.twig';
         $controllerCompletionText = '<div data-controller="sea';
         $documents->open(new Document($controllerCompletionUri, 'twig', 1, $controllerCompletionText));
+        $packageControllerCompletionUri = 'file:///workspace/templates/package_controller_completion.html.twig';
+        $packageControllerCompletionText = "{{ stimulus_controller('@symfony/ux-auto";
+        $documents->open(new Document($packageControllerCompletionUri, 'twig', 1, $packageControllerCompletionText));
         $actionCompletionUri = 'file:///workspace/templates/action_completion.html.twig';
         $actionCompletionText = '<button data-action="click->search#op';
         $documents->open(new Document($actionCompletionUri, 'twig', 1, $actionCompletionText));
         $targetCompletionUri = 'file:///workspace/templates/target_completion.html.twig';
         $targetCompletionText = '<input data-search-target="res';
         $documents->open(new Document($targetCompletionUri, 'twig', 1, $targetCompletionText));
+        $packageActionCompletionUri = 'file:///workspace/templates/package_action_completion.html.twig';
+        $packageActionCompletionText = "{{ stimulus_action('@symfony/ux-autocomplete/autocomplete', 'on";
+        $documents->open(new Document($packageActionCompletionUri, 'twig', 1, $packageActionCompletionText));
+        $packageTargetCompletionUri = 'file:///workspace/templates/package_target_completion.html.twig';
+        $packageTargetCompletionText = "{{ stimulus_target('symfony/ux-autocomplete/autocomplete', 'fi";
+        $documents->open(new Document($packageTargetCompletionUri, 'twig', 1, $packageTargetCompletionText));
         $unknownActionUri = 'file:///workspace/templates/unknown_action.html.twig';
         $unknownActionText = '<button data-action="search#missing"></button>';
         $documents->open(new Document($unknownActionUri, 'twig', 1, $unknownActionText));
         $projects = new ProjectRegistry();
         $projects->replace([$project]);
         $indexes = new StimulusIndexRegistry();
-        $indexes->forProject($project)->replace(true, new StimulusController(
-            'search',
-            '/workspace/assets/controllers/search_controller.js',
+        $indexes->forProject($project)->replace(
             true,
-            false,
-            ['open'],
-            ['input', 'results'],
-            ['url'],
-            [],
-            [],
-        ));
+            new StimulusController(
+                'search',
+                '/workspace/assets/controllers/search_controller.js',
+                true,
+                false,
+                ['open'],
+                ['input', 'results'],
+                ['url'],
+                [],
+                [],
+            ),
+            new StimulusController(
+                'symfony--ux-autocomplete--autocomplete',
+                '/workspace/vendor/symfony/ux-autocomplete/assets/dist/controller.js',
+                true,
+                true,
+                ['onChange'],
+                ['field'],
+                [],
+                [],
+                [],
+            ),
+        );
         $sourceIndexes = new StimulusSourceIndexRegistry();
         $sourceIndexes->forProject($project)->replace(
             $extractor->extract($project, new SourceDocument($controllerUri, 'javascript', $controllerText)),
@@ -119,8 +149,11 @@ final class StimulusProviderTest extends TestCase
         $codeLensProvider = new StimulusCodeLensProvider($documentResolver, $protocol, $sourceIndexes, $extractor);
 
         self::assertSame(['search'], array_column($completionProvider->complete($this->params($converter, $controllerCompletionUri, $controllerCompletionText, \strlen($controllerCompletionText))) ?? [], 'label'));
+        self::assertSame(['symfony--ux-autocomplete--autocomplete'], array_column($completionProvider->complete($this->params($converter, $packageControllerCompletionUri, $packageControllerCompletionText, \strlen($packageControllerCompletionText))) ?? [], 'label'));
         self::assertSame(['open'], array_column($completionProvider->complete($this->params($converter, $actionCompletionUri, $actionCompletionText, \strlen($actionCompletionText))) ?? [], 'label'));
         self::assertSame(['results'], array_column($completionProvider->complete($this->params($converter, $targetCompletionUri, $targetCompletionText, \strlen($targetCompletionText))) ?? [], 'label'));
+        self::assertSame(['onChange'], array_column($completionProvider->complete($this->params($converter, $packageActionCompletionUri, $packageActionCompletionText, \strlen($packageActionCompletionText))) ?? [], 'label'));
+        self::assertSame(['field'], array_column($completionProvider->complete($this->params($converter, $packageTargetCompletionUri, $packageTargetCompletionText, \strlen($packageTargetCompletionText))) ?? [], 'label'));
 
         $actionParams = $this->params($converter, $usageUri, $usageText, strpos($usageText, '#open') + 2);
         self::assertSame([$controllerUri], array_column($relationshipProvider->definition($actionParams) ?? [], 'uri'));
@@ -132,6 +165,20 @@ final class StimulusProviderTest extends TestCase
         $unknownActionParams = $this->params($converter, $unknownActionUri, $unknownActionText, strpos($unknownActionText, 'missing') + 2);
         self::assertNull($relationshipProvider->hover($unknownActionParams));
         self::assertSame([], $relationshipProvider->definition($unknownActionParams));
+
+        $packageControllerParams = $this->params($converter, $usageUri, $usageText, strpos($usageText, 'symfony/ux-autocomplete/autocomplete') + 2);
+        self::assertSame(
+            ['file:///workspace/vendor/symfony/ux-autocomplete/assets/dist/controller.js'],
+            array_column($relationshipProvider->definition($packageControllerParams) ?? [], 'uri'),
+        );
+        self::assertCount(5, $relationshipProvider->references($packageControllerParams) ?? []);
+        foreach (['onChange', 'field'] as $member) {
+            $memberParams = $this->params($converter, $usageUri, $usageText, strpos($usageText, $member) + 2);
+            self::assertSame(
+                ['file:///workspace/vendor/symfony/ux-autocomplete/assets/dist/controller.js'],
+                array_column($relationshipProvider->definition($memberParams) ?? [], 'uri'),
+            );
+        }
 
         $diagnostics = $diagnosticProvider->diagnostics(['textDocument' => ['uri' => $usageUri]]) ?? [];
         self::assertSame(['stimulus.unknown_controller'], array_column($diagnostics, 'code'));